Revert "Create maverrors.py to resolve circular imports"
This reverts commit 0b9456ef8a1aa10b98653932cdb49221aa701f0e.
